Refactor example text rendering into a common module.
