+++ examples/common/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2023 the Glazier Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//! Common functionality for Glazier examples.
+
+pub mod text;
