+++ examples/shello.rs
@@ -1,19 +1,23 @@
+use std::any::Any;
+
+use parley::FontContext;
+use tracing_subscriber::EnvFilter;
+use vello::util::{RenderContext, RenderSurface};
+use vello::Renderer;
+use vello::{
+    kurbo::{Affine, PathEl, Point, Rect},
+    peniko::{Brush, Color, Fill, Mix, Stroke},
+    RenderParams, RendererOptions, Scene, SceneBuilder,
+};
+
 use glazier::kurbo::Size;
 use glazier::{
     Application, Cursor, FileDialogToken, FileInfo, IdleToken, KeyEvent, PointerEvent, Region,
     Scalable, TimerToken, WinHandler, WindowHandle,
 };
-use parley::{FontContext, Layout};
-use std::any::Any;
-use tracing_subscriber::EnvFilter;
-use vello::util::{RenderContext, RenderSurface};
-use vello::Renderer;
-use vello::{
-    glyph::{fello::raw::FontRef, GlyphContext},
-    kurbo::{Affine, PathEl, Point, Rect},
-    peniko::{Brush, Color, Fill, Mix, Stroke},
-    RenderParams, RendererOptions, Scene, SceneBuilder,
-};
+
+mod common;
+use common::text::{self, ParleyBrush};
 
 const WIDTH: usize = 2048;
 const HEIGHT: usize = 1536;
@@ -190,53 +194,6 @@ impl WinHandler for WindowState {
     }
 }
 
-#[derive(Clone, Debug)]
-pub struct ParleyBrush(pub Brush);
-
-impl Default for ParleyBrush {
-    fn default() -> ParleyBrush {
-        ParleyBrush(Brush::Solid(Color::rgb8(0, 0, 0)))
-    }
-}
-
-impl PartialEq<ParleyBrush> for ParleyBrush {
-    fn eq(&self, _other: &ParleyBrush) -> bool {
-        true // FIXME
-    }
-}
-
-impl parley::style::Brush for ParleyBrush {}
-
-// NOTE for Glazier maintenance: If this function needs an update, keep in mind that this is copied from xilem/src/text.rs.
-pub fn render_text(builder: &mut SceneBuilder, transform: Affine, layout: &Layout<ParleyBrush>) {
-    let mut gcx = GlyphContext::new();
-    for line in layout.lines() {
-        for glyph_run in line.glyph_runs() {
-            let mut x = glyph_run.offset();
-            let y = glyph_run.baseline();
-            let run = glyph_run.run();
-            let font = run.font();
-            let font_size = run.font_size();
-            let font_ref = font.as_ref();
-            if let Ok(font_ref) = FontRef::from_index(font_ref.data, font.index()) {
-                let style = glyph_run.style();
-                let vars: [(&str, f32); 0] = [];
-                let mut gp = gcx.new_provider(&font_ref, None, font_size, false, vars);
-                for glyph in glyph_run.glyphs() {
-                    if let Some(fragment) = gp.get(glyph.id, Some(&style.brush.0)) {
-                        let gx = x + glyph.x;
-                        let gy = y - glyph.y;
-                        let xform = Affine::translate((gx as f64, gy as f64))
-                            * Affine::scale_non_uniform(1.0, -1.0);
-                        builder.append(&fragment, Some(transform * xform));
-                    }
-                    x += glyph.advance;
-                }
-            }
-        }
-    }
-}
-
 pub fn render_anim_frame(scene: &mut Scene, fcx: &mut FontContext, i: u64) {
     let mut sb = SceneBuilder::for_scene(scene);
     let rect = Rect::from_origin_size(Point::new(0.0, 0.0), (1000.0, 1000.0));
@@ -267,7 +224,7 @@ pub fn render_anim_frame(scene: &mut Scene, fcx: &mut FontContext, i: u64) {
     )));
     let mut layout = layout_builder.build();
     layout.break_all_lines(None, parley::layout::Alignment::Start);
-    render_text(&mut sb, Affine::translate((100.0, 400.0)), &layout);
+    text::render_text(&mut sb, Affine::translate((100.0, 400.0)), &layout);
 
     let th = (std::f64::consts::PI / 180.0) * (i as f64);
     let center = Point::new(500.0, 500.0);
